Switch from JSON to YAML to support Cloud-Init
Change our configuration file format from JSON over to YAML
so we can be Cloud-Init compliant.

+++ src/clr-installer/tui/disk_manual.go
@@ -20,7 +20,7 @@ type ManualPartPage struct {
 type SelectedBlockDevice struct {
 	bd        *storage.BlockDevice
 	part      *storage.BlockDevice
-	freeSpace float64
+	freeSpace uint64
 }
 
 const (
